Split BlockchainNetwork into BlockchainClientHandler and BlockchainPeerServer implementations.  catsakka ops use implicit def pattern.
